Player can only move when clicking on the grid
Grass is changed to not be below unpassable terrain.
when clicking to move it checks if you click a tile on the grass tilemap, then moves you to that tile if true and ends your turn. (does not check for collision with enemies yet)
added scripts to handle the grid checking